+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **v1.0.0**:
 - added support for TMDL (see [TMDL requirements](/README.md#tmdl-requirements))
   - Edit, validate and publish directly from VSCode to the PBI Service!
+- added Dataflow Transactions and Datasources
 
 **v0.9.9**:
 - added new commands for Datast Query Scale Out